Fix copying big-endian arrays from Numpy to IDL
The byte ordering of arrays is lost in the copy, and IDL assumes
native(?) byte ordering. This commit converts Numpy arrays to native
byte ordering if they're not already in that ordering, before copying
the buffer to IDL's memory.

N.B. This has only been tested on an x86_64 (i.e. little-endian) CPU.